+++ packages/vuikit/src/library/pagination/components/pagination--page-prev.js
@@ -5,7 +5,7 @@ export default {
   functional: true,
   props: ElementPaginationPagePrev.props,
   render (h, { data, props, parent }) {
-    const { label, expanded } = props
+    const { title, expanded } = props
 
     // related to vk-pagination node rerendering,
     // abort until the context is the right one
@@ -18,7 +18,7 @@ export default {
 
     return h(ElementPaginationPagePrev, {
       props: {
-        label,
+        title,
         expanded,
         disabled: parent.prevPage < 1
       },