+++ packages/vuikit/CHANGELOG.md
@@ -3,6 +3,7 @@
  - Export components constants for easier integration with other components.
  - Refactor VkNavbar and fix JS errors when combined with `vk-navbar-dropbar`.
  - Fixed an edge issue affecting `vk-icon` components.
+ - Rename `vk-pagination-page-*` prop `label` to `title`.
 
 ## 0.8.0 (April 15, 2018)